refactor(logic): 新しい型定義を追加し、indexファイルを更新
