+++ package/main/src/types/logic/if.ts
@@ -1,0 +1,2 @@
+// if文の型
+export type IF<C extends boolean, X, Y> = C extends true ? X : Y;
