+++ package/main/src/types/logic/index.ts
@@ -1,10 +1,18 @@
 export * from "./and";
+export * from "./capitalize";
+export * from "./first";
+export * from "./if";
 export * from "./imply";
+export * from "./isAny";
 export * from "./isBoolean";
+export * from "./isFloat";
+export * from "./length";
+export * from "./lengthOfString";
 export * from "./nand";
 export * from "./nor";
 export * from "./not";
 export * from "./or";
 export * from "./stringToArray";
+export * from "./stringToUnion";
 export * from "./xnor";
 export * from "./xor";
