+++ package/main/src/types/logic/isAny.ts
@@ -1,0 +1,2 @@
+// any型かどうかを判定する型
+export type IsAny<T> = 0 extends 1 & T ? true : false;
