+++ package/main/src/types/logic/stringToUnion.ts
@@ -1,0 +1,4 @@
+// 文字列をUnion型に変換する型
+export type StringToUnion<S extends string> = S extends `${infer F}${infer R}`
+  ? F | StringToUnion<R>
+  : never;
