+++ package/main/src/types/logic/toNumber.ts
@@ -1,0 +1,2 @@
+// number型に変換する型
+export type ToNumber<S> = S extends `${infer N extends number}` ? N : never;
